Files I have currently for OAuth
Been working on OAuth all week. It seems to be a 3-step authentication
protocol, one step to verify your program is authenticated, step two
makes sure the user is authenticated for your program. Step 3 actually
authenticates for this access and gives you a token.

Test is the file I was working in, Test2 was some code I found online
that I was testing to see how well I could get it to work, to see which
parts were applicable to us. Globalss was meant to hold global
variables, but globals is apparently a keyword in python, hence the two
s's.

As far as I can tell right now, test does the authentication for our
application correctly, but I am missing something because it returns an
error code.