I GOT IT WORKING GUYS
As the summary says - python is finally displaying webpages. Potential
references:

http://stackoverflow.com/questions/10372883/simple-flask-application-that-reads-its-content-from-a-html-file-external-styl
https://www.tutorialspoint.com/flask/flask_http_methods.htm

So basically all the pages are python hosted and can be made in whatever
we want, but we're kind of circumventing JS and just using Python for it
instead. It's a nifty system, albeit it annoying to figure out the first
time. Regardless, it works currently.

It validates by GitHub, but GitHub will lock you out if you try
incorrectly too many times. Incorrect tries reload the page, correct
redirect you to the success page. We will probably want to remove the UI
folder (since it basically just got split into templates and static
anyways) - but I'm not worry about that right now.

For future reference - all .html files should go in templates, all .css
and .js should go in static.