Lots of Stuff!
tl;dr version: I converted us over from using GitHubCommunicator (the
only file we're using) to using Navigator, DataRetriever, Analyzer, and
Authenticator. All tasks are under the class that seems obvious, and
navigator is designed to communicate between those and the client.

Specifics:
GitHubCommunicator doesn't need to be used for anything anymore. I left
it in for reference (atleast until we do a bit more work), but you can
do all testing on Navigator.

Navigator is the file you will run.

I figured out how to dynamically choose a github project. under
navigator, currently line "29". Change that to the user's name, followed
by their repo that you want to analyze. That will change it accordingly.

I made skeleton code for just about everything (putting in the functions
atleast). If you don't know what to do, work on making one of the
functions work.

I renamed the two pages we had to be "login" and "login_success" (.html
files). I felt it was better than "index" and "success". I updated all
code to reflect those changes.

It is very likely that if something returns "None", it shouldn't do
that. It just wanted a return statement, so I gave it one.

Updated alot of variables to be camelcase, and tried to use a standard
for method names. If anything seems out of place, feel free to correct
it.