Beginning Machine Learning
Started working on machine learning. As it currently stands, I can
count the number of times words show up within a group of text strings.
Meaning, I could count how many times word X shows up in each commit,
or file, or comment, etc.

Intention is to develop a list of acceptable and unacceptable words for
each programming language and have a user store these on the database.
This list of words will get searched for in each user’s code, and for
each file if an acceptable word is found, it will start thinking it is
that programming language. Unacceptable words will disqualify the
language. This will be used for determining which languages a user can
code in.