Contribution pull size fix
I figured out how to get it to pull more than 20 data per page, and how
to tell it to pull from the next page (if it exists). Meaning, it
should be able to handle any size project now. I tried it for about 300
entries of a project (so 3 pages) and it worked fine. Also tested it on
a project with 1 page (ours) and it worked fine.