Teams are done!
OK, so the overview version:

Clustering was done before. I used the data from clustering to
determine what % of the lines of code were to each cluster. I then
compared each user’s %’s with the other users (for each cluster). Any
users with a total overlap of a certain value got put together as
teammates.

Also added in a catch so that users with too little contribution don’t
get teammate considerations, as they didn’t contribute enough to
actually be considered on a team.

Actual values for both of those things may need tweaked, but the actual
code is implemented and the values are simply a variable.