Leadership Ratings
Leadership ratings are in. They’re based entirely off of files and
branches led. By led, I mean the person who committed to it the most.

Leadership and teams are returned in the data structure under
“leadership” and “teams”. The actual count of branches led and files
led are under “uniqueStats” then “filesLed” and “branchesLed”.

I’ll need you to check out some stuff and A-OK it, but I think it’s
good. I’m signing off for now.