re-Presentation ready
Alli, if you’re reading this before I’m awake, please try to run pytest
test.py . This was the commit for making it work with my stuff, but
when I tried it, it threw an error. Now I know that SOUNDS bad, but it
isn’t - the error is that I’ve been using pythonw to run my code, and
pytest doesn’t use that special formatting. It won’t work on my PC.
Hence, why I want you to try it. It works on mine if I comment out
scikit and mysql.